+++ app/src/main/AndroidManifest.xml
@@ -77,7 +77,7 @@
 
         <!-- splash -->
         <activity
-            android:name=".ui.screen.splash.SplashActivity"
+            android:name=".ui.screen.StartActivity"
             android:configChanges="keyboardHidden|orientation|screenSize"
             android:exported="true"
             android:launchMode="singleInstance"
