+++ app/src/main/java/com/tari/android/wallet/di/ApplicationComponent.kt
@@ -40,6 +40,7 @@ import com.tari.android.wallet.service.service.WalletService
 import com.tari.android.wallet.ui.common.CommonViewModel
 import com.tari.android.wallet.ui.component.clipboardController.WalletAddressViewModel
 import com.tari.android.wallet.ui.component.networkStateIndicator.ConnectionIndicatorViewModel
+import com.tari.android.wallet.ui.screen.StartActivity
 import com.tari.android.wallet.ui.screen.auth.AuthActivity
 import com.tari.android.wallet.ui.screen.auth.AuthViewModel
 import com.tari.android.wallet.ui.screen.biometrics.ChangeBiometricsViewModel
@@ -97,7 +98,6 @@ import com.tari.android.wallet.ui.screen.settings.themeSelector.ThemeSelectorVie
 import com.tari.android.wallet.ui.screen.settings.torBridges.TorBridgesSelectionViewModel
 import com.tari.android.wallet.ui.screen.settings.torBridges.customBridges.CustomTorBridgesViewModel
 import com.tari.android.wallet.ui.screen.settings.userAutorization.BiometricAuthenticationViewModel
-import com.tari.android.wallet.ui.screen.splash.SplashActivity
 import com.tari.android.wallet.ui.screen.tx.details.TxDetailsViewModel
 import com.tari.android.wallet.ui.screen.tx.details.gif.GifViewModel
 import com.tari.android.wallet.ui.screen.tx.history.all.AllTxHistoryViewModel
@@ -128,7 +128,7 @@ interface ApplicationComponent {
 
     fun inject(service: WalletService)
 
-    fun inject(activity: SplashActivity)
+    fun inject(activity: StartActivity)
     fun inject(activity: OnboardingFlowActivity)
     fun inject(activity: AuthActivity)
     fun inject(activity: HomeActivity)
