+++ app/src/main/java/com/tari/android/wallet/notification/NotificationBroadcastReceiver.kt
@@ -36,12 +36,12 @@ import android.content.BroadcastReceiver
 import android.content.Context
 import android.content.Intent
 import com.orhanobut.logger.Logger
-import com.tari.android.wallet.di.DiContainer
 import com.tari.android.wallet.data.recovery.WalletRestorationState
 import com.tari.android.wallet.data.recovery.WalletRestorationStateHandler
+import com.tari.android.wallet.di.DiContainer
+import com.tari.android.wallet.ui.screen.StartActivity
 import com.tari.android.wallet.ui.screen.home.HomeActivity
 import com.tari.android.wallet.ui.screen.restore.activity.WalletRestoreActivity
-import com.tari.android.wallet.ui.screen.splash.SplashActivity
 import javax.inject.Inject
 
 class NotificationBroadcastReceiver : BroadcastReceiver() {
@@ -65,7 +65,7 @@ class NotificationBroadcastReceiver : BroadcastReceiver() {
             if (HomeActivity.instance.get() != null) {
                 Intent(context, HomeActivity::class.java)
             } else {
-                Intent(context, SplashActivity::class.java)
+                Intent(context, StartActivity::class.java)
             }
         }
         context.startActivity(newIntent.apply {
